Refactor for foundation of properties iterators
Moves pusherators into their own subpackage `pusherator`.
Preparing to put properties in `hydroflow::properties`.
