+++ hydroflow/examples/covid_tracing/main.rs
@@ -5,8 +5,8 @@ use std::time::Duration;
 
 use hydroflow::compiled::pull::JoinState;
 use hydroflow::compiled::pull::SymmetricHashJoin;
-use hydroflow::compiled::{InputBuild, IteratorToPusherator, PusheratorBuild};
 use hydroflow::lang::collections::Iter;
+use hydroflow::pusherator::{InputBuild, IteratorToPusherator, PusheratorBuild};
 use hydroflow::scheduled::graph::Hydroflow;
 use hydroflow::scheduled::graph_ext::GraphExt;
 use hydroflow::scheduled::handoff::VecHandoff;
