+++ hydroflow/src/compiled/mod.rs
@@ -1,305 +1,3 @@
-pub mod filter;
-pub mod filter_map;
-pub mod flatten;
-pub mod for_each;
-pub mod group_by;
-pub mod map;
-pub mod partition;
-pub mod pivot;
 pub mod pull;
+
 pub mod push_handoff;
-pub mod tee;
-
-use std::marker::PhantomData;
-
-pub trait Pusherator: Sized {
-    type Item;
-    fn give(&mut self, item: Self::Item);
-}
-
-pub trait IteratorToPusherator: Iterator {
-    fn pull_to_push(self) -> pivot::PivotBuild<Self>
-    where
-        Self: Sized,
-    {
-        pivot::PivotBuild::new(self)
-    }
-}
-impl<I> IteratorToPusherator for I where I: Sized + Iterator {}
-
-pub trait PusheratorBuild {
-    type Item;
-
-    type Output<O: Pusherator<Item = Self::Item>>;
-    fn build<O>(self, input: O) -> Self::Output<O>
-    where
-        O: Pusherator<Item = Self::Item>;
-
-    fn map<U, F>(self, f: F) -> map::MapBuild<Self::Item, U, F, Self>
-    where
-        Self: Sized,
-        F: FnMut(Self::Item) -> U,
-    {
-        map::MapBuild::new(self, f)
-    }
-
-    fn filter<F>(self, f: F) -> filter::FilterBuild<Self::Item, F, Self>
-    where
-        Self: Sized,
-        F: FnMut(&Self::Item) -> bool,
-    {
-        filter::FilterBuild::new(self, f)
-    }
-
-    fn tee<O1>(self, first_out: O1) -> tee::TeeBuild<Self::Item, O1, Self>
-    where
-        Self: Sized,
-        Self::Item: Clone,
-        O1: Pusherator<Item = Self::Item>,
-    {
-        tee::TeeBuild::new(self, first_out)
-    }
-
-    fn for_each<F>(self, f: F) -> Self::Output<for_each::ForEach<Self::Item, F>>
-    where
-        Self: Sized,
-        F: FnMut(Self::Item),
-    {
-        self.build(for_each::ForEach::new(f))
-    }
-
-    fn handoff<H>(
-        self,
-        handoff: &mut crate::scheduled::port::SendCtx<H>,
-    ) -> Self::Output<push_handoff::PushHandoff<'_, H, Self::Item>>
-    where
-        Self: Sized,
-        H: crate::scheduled::handoff::Handoff + crate::scheduled::handoff::CanReceive<Self::Item>,
-    {
-        self.build(push_handoff::PushHandoff::new(handoff))
-    }
-}
-
-pub struct InputBuild<T>(PhantomData<T>);
-impl<T> Default for InputBuild<T> {
-    fn default() -> Self {
-        Self(PhantomData)
-    }
-}
-impl<T> InputBuild<T> {
-    pub fn new() -> Self {
-        Default::default()
-    }
-}
-impl<T> PusheratorBuild for InputBuild<T> {
-    type Item = T;
-
-    type Output<O: Pusherator<Item = Self::Item>> = O;
-    fn build<O>(self, input: O) -> Self::Output<O>
-    where
-        O: Pusherator<Item = Self::Item>,
-    {
-        input
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::rc::Rc;
-
-    use super::{
-        filter::Filter, for_each::ForEach, group_by::GroupBy, map::Map, partition::Partition,
-        pivot::Pivot, tee::Tee, Pusherator,
-    };
-    use crate::lang::lattice::ord::MaxRepr;
-
-    #[test]
-    fn linear_chains() {
-        let mut v = Vec::new();
-        let mut pusher = Map::new(
-            |x| x * 2,
-            Filter::new(|x| *x > 5, ForEach::new(|x| v.push(x))),
-        );
-
-        for i in 0..5 {
-            pusher.give(i);
-        }
-
-        assert_eq!(v, vec![6, 8]);
-    }
-
-    #[test]
-    fn partition() {
-        let mut evens = Vec::new();
-        let mut odds = Vec::new();
-        let mut pusher = Partition::new(
-            |x| x % 2 == 0,
-            ForEach::new(|x| evens.push(x)),
-            ForEach::new(|x| odds.push(x)),
-        );
-
-        for i in 0..5 {
-            pusher.give(i);
-        }
-
-        assert_eq!(evens, vec![0, 2, 4]);
-        assert_eq!(odds, vec![1, 3]);
-    }
-
-    #[test]
-    fn tee() {
-        let mut left = Vec::new();
-        let mut right = Vec::new();
-        let mut pusher = Tee::new(
-            ForEach::new(|x| left.push(x)),
-            ForEach::new(|x| right.push(x)),
-        );
-
-        for i in 0..5 {
-            pusher.give(i);
-        }
-
-        assert_eq!(left, vec![0, 1, 2, 3, 4]);
-        assert_eq!(right, vec![0, 1, 2, 3, 4]);
-    }
-
-    #[test]
-    fn tee_rcs() {
-        let mut left = Vec::new();
-        let mut right = Vec::new();
-        let mut pusher = Map::new(
-            Rc::new,
-            Tee::new(
-                ForEach::new(|x: Rc<i32>| left.push(*x)),
-                ForEach::new(|x: Rc<i32>| right.push(*x)),
-            ),
-        );
-
-        for i in 0..5 {
-            pusher.give(i);
-        }
-
-        assert_eq!(left, vec![0, 1, 2, 3, 4]);
-        assert_eq!(right, vec![0, 1, 2, 3, 4]);
-    }
-
-    #[test]
-    fn pivot() {
-        let a = 0..10;
-        let b = 10..20;
-
-        let mut left = Vec::new();
-        let mut right = Vec::new();
-
-        let pivot = Pivot::new(
-            a.into_iter().chain(b.into_iter()),
-            Partition::new(
-                |x| x % 2 == 0,
-                ForEach::new(|x| left.push(x)),
-                ForEach::new(|x| right.push(x)),
-            ),
-        );
-
-        pivot.run();
-
-        assert_eq!(left, vec![0, 2, 4, 6, 8, 10, 12, 14, 16, 18]);
-        assert_eq!(right, vec![1, 3, 5, 7, 9, 11, 13, 15, 17, 19]);
-    }
-
-    #[test]
-    fn map_union() {
-        let mut v = Vec::new();
-        let mut pusher =
-            <GroupBy<_, MaxRepr<usize>, MaxRepr<usize>, _>>::new(ForEach::new(|x| v.push(x)));
-
-        pusher.give((1, 1));
-        pusher.give((1, 2));
-        pusher.give((1, 1));
-
-        assert_eq!(v, vec![(1, 1), (1, 2)]);
-    }
-}
-
-#[cfg(test)]
-mod test_builder {
-    use super::*;
-
-    #[test]
-    fn test_builder_constructed() {
-        let pb = InputBuild::<usize>(PhantomData);
-        let pb = filter::FilterBuild::new(pb, |&x| 0 == x % 2);
-        let pb = map::MapBuild::new(pb, |x| x * x);
-
-        let mut output = Vec::new();
-        let mut pusherator = pb.build(for_each::ForEach::new(|x| output.push(x)));
-
-        for x in 0..10 {
-            pusherator.give(x);
-        }
-
-        assert_eq!(&[0, 4, 16, 36, 64], &*output);
-    }
-
-    #[test]
-    fn test_builder() {
-        let mut output = Vec::new();
-
-        let mut pusherator = <InputBuild<usize>>::new()
-            .filter(|&x| 0 == x % 2)
-            .map(|x| x * x)
-            .for_each(|x| output.push(x));
-
-        for x in 0..10 {
-            pusherator.give(x);
-        }
-
-        assert_eq!(&[0, 4, 16, 36, 64], &*output);
-    }
-
-    #[test]
-    fn test_builder_tee() {
-        let mut output_evn = Vec::new();
-        let mut output_odd = Vec::new();
-
-        let mut pusherator = <InputBuild<usize>>::new()
-            .tee(
-                <InputBuild<usize>>::new()
-                    .filter(|&x| 0 == x % 2)
-                    .for_each(|x| output_evn.push(x)),
-            )
-            .filter(|&x| 1 == x % 2)
-            .for_each(|x| output_odd.push(x));
-
-        for x in 0..10 {
-            pusherator.give(x);
-        }
-
-        assert_eq!(&[0, 2, 4, 6, 8], &*output_evn);
-        assert_eq!(&[1, 3, 5, 7, 9], &*output_odd);
-    }
-
-    #[test]
-    fn test_built_subgraph() {
-        let mut output_evn = Vec::new();
-        let mut output_odd = Vec::new();
-
-        let pivot = [1, 2, 3, 4, 5]
-            .into_iter()
-            .chain([3, 4, 5, 6, 7])
-            .map(|x| x * 9)
-            .pull_to_push()
-            .map(|x| if 0 == x % 2 { x / 2 } else { 3 * x + 1 })
-            .tee(
-                <InputBuild<usize>>::new()
-                    .filter(|&x| 0 == x % 2)
-                    .for_each(|x| output_evn.push(x)),
-            )
-            .filter(|&x| 1 == x % 2)
-            .for_each(|x| output_odd.push(x));
-
-        pivot.run();
-
-        assert_eq!(&[28, 82, 18, 136, 82, 18, 136, 190], &*output_evn);
-        assert_eq!(&[9, 27], &*output_odd);
-    }
-}
