+++ pusherator/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "pusherator"
+version = "0.1.0"
+edition = "2021"
