+++ pusherator/src/for_each.rs
